+++ src/Diagnostics/helicity.jl
@@ -38,30 +38,28 @@ keyword argument.
 """
 function helicity end
 
-function helicity(fs::VectorOfFilaments, vs::SetOfFilamentsData, p; kws...)  # p could be Γ::Real or a ParamsBiotSavart
-    T = number_type(fs)
-    @assert T <: AbstractFloat
-    H = zero(T)
-    for (f, v) ∈ zip(fs, vs)
-        H += helicity(f, v, p; kws...)
+function helicity(fs::VectorOfFilaments, vs::SetOfFilamentsData, p; quad = nothing, nthreads = Threads.nthreads())  # p could be Γ::Real or a ParamsBiotSavart
+    maybe_parallelise_sum(fs, nthreads) do i, inds
+        helicity(fs[i], vs[i], p; quad, inds)
     end
-    H
 end
 
-function helicity(f::AbstractFilament, vs::SingleFilamentData, Γ::Real; quad = nothing)
-    _helicity(quad, f, vs, Γ)
+function helicity(f::AbstractFilament, vs::SingleFilamentData, Γ::Real; quad = nothing, inds = eachindex(f))
+    _helicity(quad, f, vs, inds, Γ)
 end
 
 function helicity(f::AbstractFilament, vs::SingleFilamentData, p::ParamsBiotSavart; kws...)
     helicity(f, vs, p.Γ; kws...)
 end
 
-function _helicity(::Nothing, f::AbstractFilament, vs::SingleFilamentData, Γ::Real)
+function _helicity(::Nothing, f::AbstractFilament, vs::SingleFilamentData, inds, Γ::Real)
+    @assert eachindex(f) == eachindex(vs)
+    checkbounds(f, inds)
     T = number_type(f)
     @assert T <: AbstractFloat
     H = zero(T)
     ts = knots(f)
-    for i ∈ eachindex(f, vs)
+    @inbounds for i ∈ inds
         v⃗ = vs[i]
         s⃗′ = f[i, Derivative(1)]
         δt = (ts[i + 1] - ts[i - 1]) / 2
@@ -71,15 +69,17 @@ function _helicity(::Nothing, f::AbstractFilament, vs::SingleFilamentData, Γ::R
 end
 
 # Case with quadratures (requires interpolating the velocity along filaments)
-function _helicity(quad, f::AbstractFilament, vs::SingleFilamentData, Γ::Real)
-    _helicity(isinterpolable(vs), quad, f, vs, Γ)
+function _helicity(quad, f::AbstractFilament, vs::SingleFilamentData, inds, Γ::Real)
+    _helicity(isinterpolable(vs), quad, f, vs, inds, Γ)
 end
 
-function _helicity(::IsInterpolable{true}, quad, f::AbstractFilament, vs::SingleFilamentData, Γ::Real)
+function _helicity(::IsInterpolable{true}, quad, f::AbstractFilament, vs::SingleFilamentData, inds, Γ::Real)
+    @assert eachindex(f) == eachindex(vs)
+    checkbounds(f, inds)
     T = number_type(f)
     @assert T <: AbstractFloat
     H = zero(T)
-    for i ∈ eachindex(segments(f))
+    for i ∈ inds
         H += integrate(f, i, quad) do f, i, ζ
             v⃗ = vs(i, ζ)
             s⃗′ = f(i, ζ, Derivative(1))
@@ -89,7 +89,9 @@ function _helicity(::IsInterpolable{true}, quad, f::AbstractFilament, vs::Single
     T(Γ) * H
 end
 
-function _helicity(::IsInterpolable{false}, quad, f::AbstractFilament, vs::SingleFilamentData, Γ::Real)
+function _helicity(::IsInterpolable{false}, quad, f::AbstractFilament, vs::SingleFilamentData, inds, Γ::Real)
+    @assert eachindex(f) == eachindex(vs)
+    checkbounds(f, inds)
     T = number_type(f)
     @assert T <: AbstractFloat
     H = zero(T)
@@ -115,7 +117,7 @@ function _helicity(::IsInterpolable{false}, quad, f::AbstractFilament, vs::Singl
         coefs = Filaments.init_coefficients(method, cs, cderiv)
         copyto!(cs, vs)
         Filaments.compute_coefficients!(coefs, ts)
-        for i ∈ eachindex(segments(f))
+        for i ∈ inds
             H += integrate(f, i, quad) do f, i, ζ
                 v⃗ = Filaments.evaluate(coefs, ts, i, ζ)
                 s⃗′ = f(i, ζ, Derivative(1))